+++ backend/main.go
@@ -9,6 +9,7 @@ import (
 	"crawlab/model"
 	"crawlab/routes"
 	"crawlab/services"
+	"crawlab/services/challenge"
 	"crawlab/services/rpc"
 	"github.com/apex/log"
 	"github.com/gin-gonic/gin"
@@ -91,6 +92,22 @@ func main() {
 			panic(err)
 		}
 		log.Info("initialized dependency fetcher successfully")
+
+		// 初始化挑战服务
+		if err := challenge.InitChallengeService(); err != nil {
+			log.Error("init challenge service error:" + err.Error())
+			debug.PrintStack()
+			panic(err)
+		}
+		log.Info("initialized challenge service successfully")
+
+		// 初始化清理服务
+		if err := services.InitCleanService(); err != nil {
+			log.Error("init clean service error:" + err.Error())
+			debug.PrintStack()
+			panic(err)
+		}
+		log.Info("initialized clean service successfully")
 	}
 
 	// 初始化任务执行器
@@ -214,6 +231,7 @@ func main() {
 				authGroup.GET("/tasks/:id/log", routes.GetTaskLog)                          // 任务日志
 				authGroup.GET("/tasks/:id/results", routes.GetTaskResults)                  // 任务结果
 				authGroup.GET("/tasks/:id/results/download", routes.DownloadTaskResultsCsv) // 下载任务结果
+				authGroup.POST("/tasks/:id/restart", routes.RestartTask)                    // 重新开始任务
 			}
 			// 定时任务
 			{
@@ -231,6 +249,7 @@ func main() {
 				authGroup.GET("/users/:id", routes.GetUser)       // 用户详情
 				authGroup.POST("/users/:id", routes.PostUser)     // 更改用户
 				authGroup.DELETE("/users/:id", routes.DeleteUser) // 删除用户
+				authGroup.PUT("/users-add", routes.PutUser)       // 添加用户
 				authGroup.GET("/me", routes.GetMe)                // 获取自己账户
 				authGroup.POST("/me", routes.PostMe)              // 修改自己账户
 			}
@@ -254,6 +273,18 @@ func main() {
 				authGroup.POST("/projects/:id", routes.PostProject)     // 新增
 				authGroup.DELETE("/projects/:id", routes.DeleteProject) // 删除
 			}
+			// 挑战
+			{
+				authGroup.GET("/challenges", routes.GetChallengeList)          // 挑战列表
+				authGroup.POST("/challenges-check", routes.CheckChallengeList) // 检查挑战列表
+			}
+			// 操作
+			{
+				//authGroup.GET("/actions", routes.GetActionList)   // 操作列表
+				//authGroup.GET("/actions/:id", routes.GetAction)   // 操作
+				authGroup.PUT("/actions", routes.PutAction) // 新增操作
+				//authGroup.POST("/actions/:id", routes.PostAction) // 修改操作
+			}
 			// 统计数据
 			authGroup.GET("/stats/home", routes.GetHomeStats) // 首页统计数据
 			// 文件
@@ -262,7 +293,7 @@ func main() {
 			authGroup.GET("/git/branches", routes.GetGitRemoteBranches) // 获取 Git 分支
 			authGroup.GET("/git/public-key", routes.GetGitSshPublicKey) // 获取 SSH 公钥
 			authGroup.GET("/git/commits", routes.GetGitCommits)         // 获取 Git Commits
-			authGroup.POST("/git/checkout", routes.PostGitCheckout)         // 获取 Git Commits
+			authGroup.POST("/git/checkout", routes.PostGitCheckout)     // 获取 Git Commits
 		}
 	}