+++ backend/routes/config_spider.go
@@ -51,6 +51,9 @@ func PutConfigSpider(c *gin.Context) {
 	// 将FileId置空
 	spider.FileId = bson.ObjectIdHex(constants.ObjectIdNull)
 
+	// UserId
+	spider.UserId = services.GetCurrentUserId(c)
+
 	// 创建爬虫目录
 	spiderDir := filepath.Join(viper.GetString("spider.path"), spider.Name)
 	if utils.Exists(spiderDir) {
@@ -109,7 +112,11 @@ func UploadConfigSpider(c *gin.Context) {
 	spider, err := model.GetSpider(bson.ObjectIdHex(id))
 	if err != nil {
 		HandleErrorF(http.StatusBadRequest, c, fmt.Sprintf("cannot find spider (id: %s)", id))
+		return
 	}
+
+	// UserId
+	spider.UserId = services.GetCurrentUserId(c)
 
 	// 获取上传文件
 	file, header, err := c.Request.FormFile("file")
@@ -174,6 +181,7 @@ func UploadConfigSpider(c *gin.Context) {
 	// 根据序列化后的数据处理爬虫文件
 	if err := services.ProcessSpiderFilesFromConfigData(spider, configData); err != nil {
 		HandleError(http.StatusInternalServerError, c, err)
+		return
 	}
 
 	c.JSON(http.StatusOK, Response{
@@ -203,6 +211,11 @@ func PostConfigSpiderSpiderfile(c *gin.Context) {
 	if err != nil {
 		HandleErrorF(http.StatusBadRequest, c, fmt.Sprintf("cannot find spider (id: %s)", id))
 		return
+	}
+
+	// UserId
+	if !spider.UserId.Valid() {
+		spider.UserId = bson.ObjectIdHex(constants.ObjectIdNull)
 	}
 
 	// 反序列化
@@ -245,6 +258,11 @@ func PostConfigSpiderConfig(c *gin.Context) {
 	if err != nil {
 		HandleErrorF(http.StatusBadRequest, c, fmt.Sprintf("cannot find spider (id: %s)", id))
 		return
+	}
+
+	// UserId
+	if !spider.UserId.Valid() {
+		spider.UserId = bson.ObjectIdHex(constants.ObjectIdNull)
 	}
 
 	// 反序列化配置数据