+++ backend/routes/stats.go
@@ -3,6 +3,7 @@ package routes
 import (
 	"crawlab/constants"
 	"crawlab/model"
+	"crawlab/services"
 	"github.com/gin-gonic/gin"
 	"github.com/globalsign/mgo/bson"
 	"net/http"
@@ -14,6 +15,7 @@ func GetHomeStats(c *gin.Context) {
 		SpiderCount     int `json:"spider_count"`
 		ActiveNodeCount int `json:"active_node_count"`
 		ScheduleCount   int `json:"schedule_count"`
+		ProjectCount    int `json:"project_count"`
 	}
 
 	type Data struct {
@@ -22,7 +24,7 @@ func GetHomeStats(c *gin.Context) {
 	}
 
 	// 任务总数
-	taskCount, err := model.GetTaskCount(nil)
+	taskCount, err := model.GetTaskCount(bson.M{"user_id": services.GetCurrentUserId(c)})
 	if err != nil {
 		HandleError(http.StatusInternalServerError, c, err)
 		return
@@ -36,21 +38,28 @@ func GetHomeStats(c *gin.Context) {
 	}
 
 	// 爬虫总数
-	spiderCount, err := model.GetSpiderCount()
+	spiderCount, err := model.GetSpiderCount(bson.M{"user_id": services.GetCurrentUserId(c)})
 	if err != nil {
 		HandleError(http.StatusInternalServerError, c, err)
 		return
 	}
 
 	// 定时任务数
-	scheduleCount, err := model.GetScheduleCount()
+	scheduleCount, err := model.GetScheduleCount(bson.M{"user_id": services.GetCurrentUserId(c)})
+	if err != nil {
+		HandleError(http.StatusInternalServerError, c, err)
+		return
+	}
+
+	// 项目数
+	projectCount, err := model.GetProjectCount(bson.M{"user_id": services.GetCurrentUserId(c)})
 	if err != nil {
 		HandleError(http.StatusInternalServerError, c, err)
 		return
 	}
 
 	// 每日任务数
-	items, err := model.GetDailyTaskStats(bson.M{})
+	items, err := model.GetDailyTaskStats(bson.M{"user_id": services.GetCurrentUserId(c)})
 	if err != nil {
 		HandleError(http.StatusInternalServerError, c, err)
 		return
@@ -65,6 +74,7 @@ func GetHomeStats(c *gin.Context) {
 				TaskCount:       taskCount,
 				SpiderCount:     spiderCount,
 				ScheduleCount:   scheduleCount,
+				ProjectCount:    projectCount,
 			},
 			Daily: items,
 		},