+++ frontend/src/components/Config/ConfigList.vue
@@ -131,14 +131,38 @@
 
           <div class="button-group-container">
             <div class="button-group">
-              <el-button id="btn-run" size="small" type="danger" @click="onCrawl">
+              <el-button
+                id="btn-run"
+                size="small"
+                type="danger"
+                :disabled="isDisabled"
+                icon="el-icon-video-play"
+                @click="onCrawl"
+              >
                 {{$t('Run')}}
+              </el-button>
+              <el-button
+                id="btn-convert"
+                size="small"
+                type="warning"
+                :disabled="isDisabled"
+                icon="el-icon-refresh-right"
+                @click="onConvert"
+              >
+                {{$t('Convert to Customized')}}
               </el-button>
               <!--              <el-button type="primary" @click="onExtractFields" v-loading="extractFieldsLoading">-->
               <!--                {{$t('ExtractFields')}}-->
               <!--              </el-button>-->
               <!--              <el-button type="warning" @click="onPreview" v-loading="previewLoading">{{$t('Preview')}}</el-button>-->
-              <el-button id="btn-save" size="small" type="success" @click="onSave" v-loading="saveLoading">
+              <el-button
+                id="btn-save"
+                size="small"
+                type="success"
+                :disabled="saveLoading || isDisabled"
+                @click="onSave"
+                :icon="saveLoading ? 'el-icon-loading' : 'el-icon-check'"
+              >
                 {{$t('Save')}}
               </el-button>
             </div>
@@ -303,7 +327,7 @@
       <!--Setting-->
       <el-tab-pane name="settings" :label="$t('Settings')">
         <div class="actions" style="text-align: right;margin-bottom: 10px">
-          <el-button type="success" size="small" @click="onSave">
+          <el-button type="success" size="small" :disabled="isDisabled" @click="onSave">
             {{$t('Save')}}
           </el-button>
         </div>
@@ -316,7 +340,13 @@
       <!--Spiderfile-->
       <el-tab-pane name="spiderfile" label="Spiderfile">
         <div class="spiderfile-actions">
-          <el-button type="primary" size="small" style="margin-right: 10px;" @click="onSpiderfileSave">
+          <el-button
+            type="primary"
+            size="small"
+            style="margin-right: 10px;"
+            :disabled="isDisabled"
+            @click="onSpiderfileSave"
+          >
             <font-awesome-icon :icon="['fa', 'save']"/>
             {{$t('Save')}}
           </el-button>
@@ -330,7 +360,10 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
+import {
+  mapState,
+  mapGetters
+} from 'vuex'
 import echarts from 'echarts'
 import FieldsTableView from '../TableView/FieldsTableView'
 import CrawlConfirmDialog from '../Common/CrawlConfirmDialog'
@@ -501,6 +534,12 @@ export default {
       'spiderForm',
       'previewCrawlData'
     ]),
+    ...mapGetters('user', [
+      'userInfo'
+    ]),
+    isDisabled () {
+      return this.spiderForm.is_public && this.spiderForm.username !== this.userInfo.username && this.userInfo.role !== 'admin'
+    },
     fields () {
       if (this.spiderForm.crawl_type === 'list') {
         return this.spiderForm.fields
@@ -987,10 +1026,33 @@ ${f.css || f.xpath} ${f.attr ? ('(' + f.attr + ')') : ''} ${f.next_stage ? (' --
       const nextStageField = this.getNextStageField(stage)
       if (!nextStageField) return
       return this.spiderForm.config.stages[nextStageField.next_stage]
+    },
+    onConvert () {
+      this.$confirm(this.$t('Are you sure to convert this spider to customized spider?'), this.$t('Notification'), {
+        confirmButtonText: this.$t('Confirm'),
+        cancelButtonText: this.$t('Cancel'),
+        type: 'warning'
+      }).then(async () => {
+        this.spiderForm.type = 'customized'
+        this.$store.dispatch('spider/editSpider')
+          .then(res => {
+            if (!res.data.error) {
+              this.$store.commit('spider/SET_CONFIG_LIST_TS', +new Date())
+              this.$message({
+                type: 'success',
+                message: 'Converted successfully'
+              })
+            } else {
+              this.$message({
+                type: 'error',
+                message: 'Converted unsuccessfully'
+              })
+            }
+            this.$store.dispatch('spider/getSpiderData', this.spiderForm._id)
+            this.$st.sendEv('爬虫详情', '配置', '转化为自定义爬虫')
+          })
+      })
     }
-  },
-  mounted () {
-    this.activeNames = this.spiderForm.config.stages.map(stage => stage.name)
   }
 }
 </script>