+++ frontend/src/i18n/zh.js
@@ -222,6 +222,16 @@ export default {
   'Add Variable': '添加变量',
   'Copy Spider': '复制爬虫',
   'New Spider Name': '新爬虫名称',
+  'All Spiders': '所有爬虫',
+  'My Spiders': '我的爬虫',
+  'Public Spiders': '公共爬虫',
+  'Is Public': '是否公共',
+  'Owner': '所有者',
+  'Convert to Customized': '转化为自定义',
+  'Is De-Duplicated': '是否去重',
+  'Please enter de-duplicated field': '请输入去重字段',
+  'Overwrite': '覆盖',
+  'Ignore': '忽略',
 
   // 爬虫列表
   'Name': '名称',
@@ -256,6 +266,9 @@ export default {
   'Empty results': '空结果',
   'Navigate to Spider': '导航到爬虫',
   'Navigate to Node': '导航到节点',
+  'Restart': '重新运行',
+  'Redirect to task detail': '跳转到任务详情页',
+  'Retry (Maximum 5 Times)': '是否重试（最多 5 次）',
 
   // 任务列表
   'Node': '节点',
@@ -269,6 +282,7 @@ export default {
 
   // 项目
   'All Tags': '全部标签',
+  'Projects': '项目',
   'Project Name': '项目名称',
   'Project Description': '项目描述',
   'Tags': '标签',
@@ -291,6 +305,7 @@ export default {
   'Cron': 'Cron',
   'Cron Expression': 'Cron 表达式',
   'Cron expression is invalid': 'Cron 表达式不正确',
+  'View Tasks': '查看任务',
 
   // 网站
   'Site': '网站',
@@ -305,6 +320,13 @@ export default {
   'Robots Protocol': 'Robots 协议',
   'Home Page Response Time (sec)': '首页响应时间(秒)',
   'Home Page Response Status Code': '首页响应状态码',
+
+  // 反馈
+  'Feedback': '反馈',
+  'Feedbacks': '反馈',
+  'Wechat': '微信',
+  'Content': '内容',
+  'Rating': '评分',
 
   // 用户
   'Super Admin': '超级管理员',
@@ -361,7 +383,11 @@ export default {
   'Are you sure to delete this node?': '你确定要删除该节点?',
   'Are you sure to run this spider?': '你确定要运行该爬虫?',
   'Are you sure to delete this file/directory?': '你确定要删除该文件/文件夹?',
+  'Are you sure to convert this spider to customized spider?': '你确定要转化该爬虫为自定义爬虫?',
+  'Are you sure to delete this task?': '您确定要删除该任务?',
   'Added spider successfully': '成功添加爬虫',
+  'Converted successfully': '成功转化',
+  'Converted unsuccessfully': '未成功转化',
   'Uploaded spider files successfully': '成功上传爬虫文件',
   'Node info has been saved successfully': '节点信息已成功保存',
   'A task has been scheduled successfully': '已经成功派发一个任务',
@@ -419,6 +445,7 @@ export default {
   'How to Upgrade': '升级方式',
   'Release': '发布',
   'Add Wechat to join discussion group': '添加微信 tikazyq1 加入交流群',
+  'Submitted successfully': '提交成功',
 
   // 登录
   'Sign in': '登录',
@@ -458,6 +485,14 @@ export default {
   'Allow Sending Statistics': '允许发送统计信息',
   'General': '通用',
   'Enable Tutorial': '启用教程',
+
+  // 挑战
+  'Challenge': '挑战',
+  'Challenges': '挑战',
+  'Difficulty': '难度',
+  'Achieved': '已达成',
+  'Not Achieved': '未达成',
+  'Start Challenge': '开始挑战',
 
   // 全局
   'Related Documentation': '相关文档',
@@ -584,6 +619,11 @@ docker run -d --restart always --name crawlab_worker \\
   'Are you sure to stop selected items?': '您是否确认停止所选项？',
   'Sent signals to cancel selected tasks': '已经向所选任务发送取消任务信号',
   'Copied successfully': '已成功复制',
+  'You have started the challenge.': '您已开始挑战',
+  'Please enter your email': '请输入您的邮箱',
+  'Please enter your Wechat account': '请输入您的微信账号',
+  'Please enter your feedback content': '请输入您的反馈内容',
+  'No response from the server. Please make sure your server is running correctly. You can also refer to the documentation to solve this issue.': '服务器无响应，请保证您的服务器正常运行。您也可以参考文档来解决这个问题（文档链接在下方）',
 
   // 其他
   'Star crawlab-team/crawlab on GitHub': '在 GitHub 上为 Crawlab 加星吧'